+++ sdk/core/azure-core/src/main/java/com/azure/core/implementation/serializer/HttpResponseHeaderDecoder.java
@@ -4,6 +4,7 @@
 package com.azure.core.implementation.serializer;
 
 import com.azure.core.exception.HttpResponseException;
+import com.azure.core.http.HttpHeaders;
 import com.azure.core.http.HttpResponse;
 import com.azure.core.util.serializer.SerializerAdapter;
 import reactor.core.Exceptions;
@@ -12,31 +13,30 @@ import java.io.IOException;
 import java.lang.reflect.Type;
 
 /**
- * Decoder to decode header of HTTP response.
+ * {@link HttpResponse} {@link HttpHeaders} decoder.
  */
 final class HttpResponseHeaderDecoder {
     private static final String MALFORMED_HEADERS_MESSAGE = "HTTP response has malformed headers";
 
     /**
-     * Decode headers of the http response.
+     * Decodes the {@link HttpHeaders} in an {@link HttpResponse}.
+     * <p>
+     * If there is no decoded headers type for the API that was called null will be returned.
      *
-     * The decoding happens when caller subscribed to the returned {@code Mono<Object>}, if the response header is not
-     * decodable then {@code Mono.empty()} will be returned.
-     *
-     * @param response the response containing the headers to be decoded
-     * @param serializer the adapter to use for decoding
-     * @param decodeData the necessary data required to decode a Http response
-     * @return publisher that emits decoded response header upon subscription if header is decodable, no emission if the
-     * header is not-decodable
+     * @param response The response containing the headers to decode.
+     * @param serializer The adapter that will perform decoding.
+     * @param decodedHeadersType The decoded headers type.
+     * @return The decoded headers object, or null if {@code decodedHeadersType} is null.
+     * @throws HttpResponseException If the response headers fail to decode to the decoded headers type.
      */
-    static Object decode(HttpResponse response, SerializerAdapter serializer, HttpResponseDecodeData decodeData) {
-        Type headerType = decodeData.getHeadersType();
-        if (headerType == null) {
+    static Object decode(HttpResponse response, SerializerAdapter serializer, Type decodedHeadersType) {
+        // There is no type to decode into, return null.
+        if (decodedHeadersType == null) {
             return null;
         }
 
         try {
-            return serializer.deserialize(response.getHeaders(), headerType);
+            return serializer.deserialize(response.getHeaders(), decodedHeadersType);
         } catch (IOException ex) {
             throw Exceptions.propagate(new HttpResponseException(MALFORMED_HEADERS_MESSAGE, response, ex));
         }