+++ sdk/spring/spring-cloud-azure-stream-binder-eventhubs-core/src/main/java/com/azure/spring/cloud/stream/binder/eventhubs/core/properties/EventHubsExtendedBindingProperties.java
@@ -41,6 +41,7 @@ public class EventHubsExtendedBindingProperties
      *
      * @return Bindings value
      */
+    @Override
     public Map<String, EventHubsBindingProperties> getBindings() {
         return doGetBindings();
     }